code/simple_examples/requirements.txt
canvasapi>=3.2,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
markdown>=3.5.1,<4.0
markdownify>=0.11.6,<0.12
html2text>=2020.1.16,<2021.0
